+++ @geogebra/solver-sdk/src/renderer/index.ts
@@ -1,5 +1,5 @@
-export type { LatexSettings, LatexTransformer } from './tree-to-latex';
-export { treeToLatex } from './tree-to-latex';
+export type { LatexSettings, LatexTransformer, MathWords } from './tree-to-latex';
+export { treeToLatex, defaultMathWords } from './tree-to-latex';
 export { setsSolutionFormatter, simpleSolutionFormatter } from './solution-formatter';
 export { treeToSolver } from './tree-to-solver';
 export { specialTransformationLatex } from './transformation-to-latex';